feat: rename 'models' package to a meaningful one 'storage'
