+++ server/controllers/users.go
@@ -7,9 +7,9 @@ import (
 	jwt "github.com/dgrijalva/jwt-go"
 	"github.com/labstack/echo"
 	log "github.com/sirupsen/logrus"
-	"github.com/soprasteria/docktor/server/models"
 	"github.com/soprasteria/docktor/server/modules/auth"
 	"github.com/soprasteria/docktor/server/modules/users"
+	"github.com/soprasteria/docktor/server/storage"
 	"github.com/soprasteria/docktor/server/types"
 	"gopkg.in/mgo.v2/bson"
 )
@@ -20,7 +20,7 @@ type Users struct {
 
 //GetAll users from docktor
 func (u *Users) GetAll(c echo.Context) error {
-	docktorAPI := c.Get("api").(*models.Docktor)
+	docktorAPI := c.Get("api").(*storage.Docktor)
 	webservice := users.Rest{Docktor: docktorAPI}
 	users, err := webservice.GetAllUserRest()
 	if err != nil {
@@ -33,7 +33,7 @@ func (u *Users) GetAll(c echo.Context) error {
 // Update user into docktor
 // Only admin and current user is able to update a user
 func (u *Users) Update(c echo.Context) error {
-	docktorAPI := c.Get("api").(*models.Docktor)
+	docktorAPI := c.Get("api").(*storage.Docktor)
 	authenticatedUser, err := getUserFromToken(c)
 	if err != nil {
 		return c.String(http.StatusUnauthorized, auth.ErrInvalidCredentials.Error())
@@ -108,7 +108,7 @@ func (u *Users) Update(c echo.Context) error {
 
 //Delete user into docktor
 func (u *Users) Delete(c echo.Context) error {
-	docktorAPI := c.Get("api").(*models.Docktor)
+	docktorAPI := c.Get("api").(*storage.Docktor)
 	id := c.Param("userID")
 
 	authenticatedUser, err := getUserFromToken(c)
@@ -170,7 +170,7 @@ func (u *Users) ChangePassword(c echo.Context) error {
 		return c.String(http.StatusForbidden, "New password should not be empty and be at least 6 characters")
 	}
 
-	docktorAPI := c.Get("api").(*models.Docktor)
+	docktorAPI := c.Get("api").(*storage.Docktor)
 	webservice := auth.Authentication{Docktor: docktorAPI}
 	err = webservice.ChangePassword(authenticatedUser.ID, options.OldPassword, options.NewPassword)
 
@@ -204,7 +204,7 @@ func (u *Users) Get(c echo.Context) error {
 }
 
 func getUserFromToken(c echo.Context) (users.UserRest, error) {
-	docktorAPI := c.Get("api").(*models.Docktor)
+	docktorAPI := c.Get("api").(*storage.Docktor)
 	userToken := c.Get("user-token").(*jwt.Token)
 
 	claims := userToken.Claims.(*auth.MyCustomClaims)
