+++ server/modules/daemons/middlewares.go
@@ -5,13 +5,13 @@ import (
 	"net/http"
 
 	"github.com/labstack/echo"
-	"github.com/soprasteria/docktor/server/models"
+	"github.com/soprasteria/docktor/server/storage"
 )
 
 // RetrieveDaemon find daemon using id param and put it in echo.Context
 func RetrieveDaemon(next echo.HandlerFunc) echo.HandlerFunc {
 	return func(c echo.Context) error {
-		docktorAPI := c.Get("api").(*models.Docktor)
+		docktorAPI := c.Get("api").(*storage.Docktor)
 		daemonID := c.Param("daemonID")
 		if daemonID == "" {
 			return c.String(http.StatusBadRequest, DaemonInvalidID)
