+++ server/modules/services/middlewares.go
@@ -5,13 +5,13 @@ import (
 	"net/http"
 
 	"github.com/labstack/echo"
-	"github.com/soprasteria/docktor/server/models"
+	"github.com/soprasteria/docktor/server/storage"
 )
 
 // RetrieveService find service using id param and put it in echo.Context
 func RetrieveService(next echo.HandlerFunc) echo.HandlerFunc {
 	return func(c echo.Context) error {
-		docktorAPI := c.Get("api").(*models.Docktor)
+		docktorAPI := c.Get("api").(*storage.Docktor)
 		serviceID := c.Param("serviceID")
 		if serviceID == "" {
 			return c.String(http.StatusBadRequest, ServiceInvalidID)
