+++ server/modules/users/middlewares.go
@@ -5,13 +5,13 @@ import (
 	"net/http"
 
 	"github.com/labstack/echo"
-	"github.com/soprasteria/docktor/server/models"
+	"github.com/soprasteria/docktor/server/storage"
 )
 
 // RetrieveUser find user using id param and put it in echo.Context
 func RetrieveUser(next echo.HandlerFunc) echo.HandlerFunc {
 	return func(c echo.Context) error {
-		docktorAPI := c.Get("api").(*models.Docktor)
+		docktorAPI := c.Get("api").(*storage.Docktor)
 		id := c.Param("userID")
 		if id == "" {
 			return c.String(http.StatusBadRequest, UserInvalidID)
