+++ server/storage/groups.go
@@ -1,4 +1,4 @@
-package models
+package storage
 
 import (
 	"fmt"
